Add __init__.py for easier eeVR installation process
