+++ VRRenderer.py
@@ -793,58 +793,3 @@ class RenderToolsPanel(Panel):
         if not context.scene.cancelVRRenderer:
             col.operator("wl.render_cancel", text="Cancel")
             col.label(text="Rendering frame {}".format(bpy.context.scene.frame_current))
- 
- 
-# Register all classes
-def register():
-    bpy.types.Scene.renderModeEnum = bpy.props.EnumProperty(
-        items = [('EQUI', 'Equirectangular', 'Renders in equirectangular projection'),
-                 ('DOME', 'Full Dome', 'Renders in full dome projection')],
-        default='EQUI',
-        name="Mode")
-    bpy.types.Scene.domeModeEnum = bpy.props.EnumProperty(
-        items = [('0', 'Equidistant (VTA)', 'Renders in equidistant dome projection'),
-                 ('1', 'Hemispherical (VTH)', 'Renders in hemispherical dome projection'),
-                 ('2', 'Equisolid', 'Renders in equisolid dome projection'),
-                 ('3', 'Stereographic', 'Renders in Stereographic dome projection')],
-        default='0',
-        name="Method")
-    bpy.types.Scene.renderFOV = bpy.props.FloatProperty(180.0,default=180.0, name="FOV", min=180, max=360,
-                                description="Field of view in degrees")
-    bpy.types.Scene.cancelVRRenderer = bpy.props.BoolProperty(name="Cancel", default=True)
-    bpy.utils.register_class(RenderImage)
-    bpy.utils.register_class(RenderAnimation)
-    bpy.utils.register_class(RenderToolsPanel)
-    bpy.utils.register_class(VRRendererCancel)
-   
-   
- 
- 
-# Unregister all classes
-def unregister():
-    del bpy.types.Scene.domeModeEnum
-    del bpy.types.Scene.renderModeEnum
-    del bpy.types.Scene.renderFOV
-    bpy.utils.unregister_class(RenderImage)
-    bpy.utils.unregister_class(RenderAnimation)
-    bpy.utils.unregister_class(RenderToolsPanel)
-    bpy.utils.unregister_class(VRRendererCancel)
- 
- 
-bl_info = {
-    "name": "eeVR",
-    "description": "Render in different projections using Eevee engine",
-    "author": "EternalTrail",
-    "version": (0, 1),
-    "blender": (2, 82, 7),
-    "location": "View3D > UI",
-    "warning": "This addon is still in early alpha, may break your blend file!",
-    "wiki_url": "https://github.com/EternalTrail/eeVR",
-    "tracker_url": "https://github.com/EternalTrail/eeVR/issues",
-    "support": "TESTING",
-    "category": "Render",
-}
- 
-# If the script is not an addon when it is run, register the classes
-if __name__=="__main__":
-    register()
